Fix languageAndLexicalCategoryWidget whitespace for Vue 3
In pure Vue 3 (but not in the migration build as used in MediaWiki core
at the moment), some more whitespace inside this template gets trimmed
by Vue, resulting in the “language” text and language link being
directly adjacent with no space in between (“LanguageEnglish”).
Rearrange the tags so that the space between these two spans doesn’t
contain a newline, which signals to Vue that this is a significant space
that shouldn’t be trimmed.

I don’t actually think this is what the markup should look like in the
long run – it should probably be an i18n message, maybe colon-separator?
– but it maintains the same behavior across the Vue 3 migration, which
should be good enough for now.

Bug: T304534
Change-Id: I8fa6511ed2aebd8aea7f987330ba58ba8325f4de