Add more logging to unexpected exception being thrown
Lexeme::getForm never returns `null`, it just throws an
OutOfRangeException if the form does not exist.

This change is similar to Ic3356de8, but does not rethrow a
PatcherException, because that exception extends RuntimeException,
whereas this problem is one of programming logic: when a Form is
removed, we should never ever end up in that code path. So the
LogicException OutOfRangeException feels more appropriate.

Bug: T326768
Change-Id: Ia99f1b883216387dba02bfb0c3e029e40695e8b6