Remove unused icomefromthenet/reverse-regex dependency
This was added in change I70d36bb963 (commit b4e3db9208), but ended up
not being used in the tests added there, at least not in the version
that was ultimately merged. (I haven’t checked if it was used in some
intermediate patch sets.) It causes two warnings about deprecated
packages on `composer install`, so let’s remove it.

Change-Id: I7f38c39161c52822902a0d9ea99d102abb52b334