Remove tests relying on foreign entity IDs
This is needed in order to remove support for them from Wikibase.

In LexemeLinkFormatterTest, I don’t think the test was ever supposed to
use foreign entity IDs (the important part of the test, if I understand
correctly, is the UnresolvedEntityRedirectException), it just happened
to work so far by accident.

Bug: T291823
Bug: T338223
Change-Id: I0a66c7a0e6f3a227d81b98635f91c7727086715c
Needed-By: I46b7eff7cfe3b472b58e9a45f2f217dc70fafbe7