Reenable and adjust API rate-limiting tests for new message
The message changed in context of T310476. But since the particular
change that introduced the issue was marked as risky, the decision was
to disable the tests for the time being, and reenable and adjust them
when that core change is known to be stable in production. That is now.

Bug: T342527
Change-Id: I2037023804bb9c6b330e10ae934ec9919427b473