Disable running the daily selenium tests against beta
In I3f08635e40 (commit d0d69d3) the original selenium-test script was
disabled by prefixing it with an @-character. This was done in order to
be update the CI job that runs these tests to a newer Node version than
Node 14 (see T293081 and T314470).
The `selenium-daily` script uses that now disabled `selenium-test`
script and has been failing since the above mentioned change was merged.

This change disables it in the same way. It should be reenabled when the
selenium tests have been either migrated to a newer version or otherwise
been modernized in a way that they can reliably run in CI again.

Bug: T293081
Change-Id: I52e8546ddab438f7bd0096d79751cbbc719965b4
Co-Authored-By: Noa Rave <noa.rave@wikimedia.de>