Actually disable running the daily selenium tests against beta
This was already attempted in I52e8546 (commit fc4588a), but the chosen
solution did not work as expected. We still got a CI error, this time
about the script `selenium-daily` missing. Apparently, the CI jobs that
executes this, does not check for the script existing, unlike the job
that runs plain `selenium-test`.

The goal is to just disable this for now in our code base, and use
T344140 to decide how we want to move forward. Then we can update the CI
infrastructure accordingly.

Bug: T293081
Bug: T344140
Change-Id: I2c7fb9ed4e81f1d980e5e5fbfc85d50031b56a60