Fix Minvera style issue by extracting Vector-specific styles
For users on "Desktop" Wikidata (www.wikidata.org) that are actively
using the Minerva skin, the top of the Lexeme header overlaps with the
toolbar. This negative margin was introduced in Ib66b5662 (commit
aed0019) for the desktop skin.

Notably, all these styles are NOT loaded for mobile Wikidata
(m.wikidata.org) at all, so it only affects people using the Desktop
version of Wikidata and the Minerva skin.

Nonetheless, this negative margin is fundamentally a workaround for some
vector-specific oddities, so I'm moving it to a vector-specific set of
styles.

This negative margin is needed both for the legacy vector and vector-2022
albeit it is different elements that are being worked around.
Legacy vector has empty `#contentSub2` and `#mw-content-subtitle` elements with a bottom-margin
of 1.4em.
Vector-2022 has its own assortment of elements with margins.

The skinStyles property of ResourceLoader modules is unfortunately only really documented in the
source code (and the docs generated from it). See MainConfigSchema.php

Bug: T318981
Change-Id: I3e51c039aba650bcf5961ccda4dca24c3e36df87