Don’t use LanguageNameLookup service in formatter callbacks
The service is deprecated, and a LanguageNameLookupFactory should be
used instead. In entitytypes, we have a proper $language we can pass
into the factory (which is more correct anyway, though I’m not sure if
this is ever different from the user language in practice); in
datatypes, we’re still using the user language (which is also
deprecated), but at least we can get rid of the LanguageNameLookup
service and leave the UserLanguage service for later.

Also, extract $services into a variable in entitytypes (and pass it into
the service getters, instead of each getter calling getInstance()
again), and add a TODO comment in datatypes.

Bug: T281721
Change-Id: I40cc870439287f6f1bdbbf3bfee8a52d79a35cd3