Replace remaining ILocalizedException::getMessageObject calls
The structure of an ApiUsageException is guaranteed to be like
this. The extra complexity of the getMessageObject method is not
necessarily needed.

In some situations we don't need a message at all but can do the
relevant assertions directly with the status object. Unfortunately
this is not always possible. Sometimes the status says "rawmessage",
which is obviously not a meaningful assertion.

Specifically, you see me removing some "modification-failed" API
code assertions. This is intentional. This string doesn't mean much
but is just some "catch all" for all kinds of errors. The code we
find in the status object is much more detailled.

Bug: T287405
Change-Id: I9c6e553542b0317323fa59497dc19ad5a87e071e