Make use of upstream assertStatusError in tests
We don't need the complexity of turning the exception into an actual,
localized Message object just to compare what is effectively the
status code. There are even upstream helper methods for this.

Note we need to turn some of the tests into
MediaWikiIntegrationTestCase to be able to use the upstream methods.
This is necessary no matter what: These tests create localized
Message objects and are in fact integrating with MediaWiki because
of this.

One of the benefits is that the upstream methods produce much more
helpful debug output.

Bug: T287405
Change-Id: I6484642abb80242de83819d955f510bab700b6e8