LexemeId, LexemeSubEntityId: Assert valid serialization in unserialize()
See the corresponding Wikibase change Id8239ce28a.

Note that FormId and SenseId, unlike LexemeId, don’t uppercase the
serialization in the constructor, so we can use the same data provider
of invalid serializations for testing the constructor and unserialize().
(We still check this->serialization !== $data['serialization'] in
LexemeSubEntityId just in case, but there should be no way that this
comparison can ever be false for those two classes.)

Change-Id: I4875b2b1b5eca49ab7f848f10563f201488a6480