Revert "Page properties should always be strings"
This was breaking our monitoring by making the pp_sortkey always null
rather than the number of statements, forms and senses.

The MediaWiki effort that originally motivated the change, making all
page properties into strings (T305158), appears to have stalled –
there’s been no activity on it since my comment in April 2022. (In
hindsight, that comment came frustratingly close to realizing this
issue, but in the end we only noticed it 19 months later.) If it ever
gets unstuck, we can consider this again, but we’ll need some solution
to keep the numeric pp_sortkey.

This reverts commit 253ede8d3ff6909c7fb8024f144fd2934b0e6cb0.

Bug: T350224
Change-Id: I5d7cb7f61ff1d04b60df6429326e59531786df59