Add update.php maintenance script to fix pp_sortkey
In production, we’ll run this manually; other installations will get it
via their update.php. (For an earlier version, not yet integrated with
update.php, see change Ia1ab07da72 in the Wikidata.org extension.)

The script gets a batch of page IDs from a replica, then sets their
value on the primary, until there are no more affected page IDs. The
replica query should be very fast, using the pp_propname_sortkey_page
index; the main reason for doing this in a separate query at all is that
I couldn’t figure out how to elegantly use a subquery in the UPDATE
WHERE clause, but I guess doing this query on the replica, even if it’s
expected to be fast, is still nice.

Bug: T350224
Change-Id: I94ad711c0408979dffbb681552f4b8cf798220b3