Remove wdio-related dependencies for the time being
    npm r wdio-{mediawiki,wikibase} @wdio/{cli,{dot,junit}-reporter,local-runner,mocha-framework}

They are not used since we disabled the wdio tests in package.json (due
to Node 16 dropping support for @wdio/sync, which we used), and cause
noise during routine package upgrades. Let’s remove them for now; if we
decide to still use WebdriverIO for browser testing (decision pending in
T346260), we can restore them.

Bug: T351383
Change-Id: Ia45cc1ded0820ca6ce1235fbd69002a8de908c98