Update structure in README-dev.md
The indentation of the table of contents and the actual heading
structure were out of sync; I’m not sure how it was intended, but I
think it makes sense to have “Development setup” as a top-level h1, and
“Prerequisites”, “Setup” and “New Lexeme Special Page” on the same level
(h2) as “Running tests”.

Remove the manually maintained table of contents. It’s redundant on
GitHub, where a real table of contents can be shown automatically; and
it didn’t fully work on Gitiles due to some anchor IDs (the numbered
ones under “Setup”) being generated differently. This leaves Gitiles
without any table of contents (there’s no auto-generated one there), but
I think the file isn’t long enough to strictly require one.

Bug: T351736
Change-Id: I620a2f8d18460f2988a7e4f7573d4656e757e567