Update unexpected-dom and add note about jsdom
While it still passes in CI with node 16, the changelog for jsdom 23
declares it to require node 18.
See: https://github.com/jsdom/jsdom/blob/main/Changelog.md#2300

So in order to not get in trouble if they do introduce code that
requires node 18 in a future minor or patch version, we should hold off
with upgrading until T331180 is done.

Bug: T352522
Change-Id: I705da59b24ec340758b6de6cf46df0c3f8383433