Remove LexemeAuthorizer, directly use EntityPermissionChecker
LexemeAuthorizer is an extra abstraction layer that we don’t need, and
by trying to be separate from MediaWiki, it requires other code to use
bad patterns (MediaWikiLexemeAuthorizer had to take the User as a
constructor parameter instead of a method parameter, which required
accessing the global request context in WikibaseLexemeServices). Remove
it, and instead have MergeLexemesInteractor directly use the underlying
EntityPermissionChecker; the tests that used SucceedingLexemeAuthorizer
and FailingLexemeAuthorizer can just as well create a mock
EntityPermissionChecker. And this way, MergeLexemesInteractor can pass
the right User, based on the IContextSource it already has, into the
EntityPermissionChecker.

Bug: T287945
Change-Id: I8bfb95f45ed85740758c04c5cc71a4e9b5d4d281
(cherry picked from commit fc476af7a76ca7aa82b5e8cf868d447a35035755)