Check entity type in Form and Sense formatters
The Form and Sense formatters assume the Entities supplied to
formatEntityId will be Forms and Senses respectively, but it is
possible to pass generic entities through the wbformatvalue API
call. Add instanceof checks here to ensure a valid entity is passed.
Otherwise, InvalidArgumentError is thrown, which API clients receive
as a generic API error.

Bug: T339904
Change-Id: I69dc6d9e45d32fcd1267f1d8f0c732a5c2139f8e