Dependency injection for temp users on merge
Pass a TempUserCreator to the MediaWikiLexemeRedirector’s parent
constructor (EntityRedirectCreationInteractor), so that it can create a
temporary user if necessary.

Note that calling methods with extra arguments is allowed in PHP, so
this change can be merged before the parent constructor actually takes
the extra argument. (But we temporarily need to silence Phan’s warning
about it.)

Bug: T356149
Change-Id: If5cc2cf1d5985e08933d1477ba857a5372fe2d1e