Correctly support temp users when merging lexemes
The underlying interactors already create temp users, but we have to
pass the right context between them; while we’re doing that, also return
the saved temp user and redirect URL from the API, and have the special
page return the redirect, with a new “success” URL state that the
redirect can return to and show the same message as before. Most of this
is based on Wikibase change I2eeca7e883.

In MergeLexemesRedirector, we now ignore LexemeRedirector::redirect()
and instead call createRedirect() which MediaWikiLexemeRedirector
inherits from EntityRedirectCreationInteractor, so that we can get the
returned context and savedTempUser; I intend to remove the now-unused
LexemeRedirector interface in a separate commit (change I77f04f7687).

Bug: T356147
Bug: T357799
Change-Id: Id9ae7ce917f39bfb105b5ff345ef361386f2d1d7
Depends-On: I2eeca7e88360eac42ee14c18814287ea08d34fa6