Use EditEntity for MergeLexemesInteractor
This lets EditEntity take care of various checks and other concerns, and
will be especially useful for IP Masking (EditEntity already knows how
to create temp users); compare Wikibase change I919db202f9. The code for
the bot flag, including the TODO about EDIT_IGNORE_CONSTRAINTS, is taken
from MediaWikiLexemeRepository; once we change the getLexeme() calls in
the interactor to use an entity (revision?) lookup, LexemeRepository
will be completely unused and we’ll be able to remove it, but let’s do
that in separate changes.

MergeLexemesInteractorTest is turned into an integration test (I’m not
sure it was 100% unit before anyway), using the real entity store,
lookup and EditEntity factory.

Bug: T356149
Bug: T356764
Change-Id: Iae0c7c3b979118559c9ce2276618c6cdec11e63d