Process CentralAuth redirects in Lexeme View for TempUsers
With T357024, the API will return a URL initialized by CentralAuth
when a TempUser tries to edit a term for the first time. We need
to redirect to this URL instead of returning to the non-edit state.

The EntityView changes are already addressed in T357149 and the
patch that this change depends on. The changes for Sense and Form
editing are addressed here.

Depends-On: Ib9afba1e239f621a5a74262e6a80a232e10f9a3f
Bug: T358323
Change-Id: I34457f9220463d2718c5f65a9db4fbfb68e4e786