Add temp user redirect parameters to API modules
In each module, use the ApiCreateTempUserTrait, add the params to
getAllowedParams(), inject an ApiHelperFactory so that we can create a
ResultBuilder in addition to the existing ApiErrorReporter, and use the
ResultBuilder to add the temp user to the response if necessary. While
we’re at it, also use the ResultBuilder’s methods markSuccess() and
addRevisionIdFromStatusToResult(), which (as far as I can tell) do the
same thing that we used to do manually; and remove the TODOs about
whether the success field is useful or not (it’s part of the stable
interface now, so if we think it’s not useful we might as well remove it
in all of Wikibase IMHO).

Bug: T357799
Change-Id: I29c056a05cd751830e7cd899dcebfffa3b3be61c