Remove LexemeRepository, use EntityRevisionLookup
This effectively reverts change I6b8bfd6bda (commit 15015057de). In my
opinion, LexemeRepository is another abstraction that didn’t turn out to
be useful, similar to LexemeRedirector (removed in change I77f04f7687,
commit d21240488c) and LexemeAuthorizer (removed in change I8bfb95f45e,
commit fc476af7a7). We never used it for anything other than merging
(regular edits, e.g. in AddForm, always used an EntityRevisionLookup for
loading and an EditEntityFactory for saving); we stopped using it for
saving merges in change Iae0c7c3b97 (since we found that
EditEntityFactory is better at its job); and the loading code can be
inlined into MergeLexemesInteractor::getLexeme() without issue, and with
the benefit of needing less catching and rethrowing of exceptions.

Change-Id: I3cc3048ef6ce4cb4634b8aafb653ca219a35bc89