tests: Fix `assert.expect(N)` and enable wikimedia/qunit rule
The wikimedia/qunit preset (and the eslint-plugin-qunit package) both
enforce the "require-expect=never-except-zero" rule by default, but
this repo had the rule disabled due to old code still relying on
manual promise and async tracking.

Remove redundant async() and expect(N) in favour returning a promise
and letting QUnit wait for it. This has the benefit of making errors
visible in QUnit output, instead of merely being an unfinished test
with "4 missing assertions".

Remove redundant async() in catch()/fail() for basic instanceof
and indexOf checks, in favour of assert.rejects()

Remove redundant async and expect(N) for non-trivial assertions
after catch(), in favour of async-await which naturally waits and
then the assertions can be done directly, thus guruanteeing they
won't be missed unless another failure is detected first.

https://qunitjs.com/api/assert/expect/
https://qunitjs.com/api/assert/rejects/

This is one of only 3 WMF-deployed repos that still using expect(N):
https://codesearch.wmcloud.org/deployed/?q=assert.expect%5C%28%5Cs%2A%5B%5E0%20%5D

Change-Id: I386edf84a75d462d78330c49e8e3a985fc1597c5