Use LogicException in MediaWikiPageSubEntityMetaDataAccessor
The BadMethodCallException is described as "Exception thrown if a
callback refers to an undefined method or if some arguments are
missing." That's not what's going on here.

I keep seeing this issue in all kinds of codebases. I believe it's
the rather generic name "bad method call" that makes people think
this can be used for all kinds of "bad" calls, whatever "bad" means.
While (mis)using an exception like this usually doesn't cause issues
I still find it unnecessarily confusing.

Change-Id: Ifedbf9c66cacdc062a683014d0a2d74b2026cc9e