Enable eslint checks for cypress files
Add eslint-typescript support to the project and add an additional
`test:` task to `npm run test` to run `eslint` over the `cypress`
specs and pageobjects.

This implementation extends the existing `grunt` task and includes
the `wmde/eslint-config-wikimedia-typescript` rules.

`cypress.config.ts` is excluded here. The project is configured as
a CommonJS project, so Javascript files at the top-level are not
allowed to include `import` or `export` statements. Changing the
project to `"type": "module"` would allow that, but confuses `grunt`
(which expects and includes CommonJS files). Cypress manages to
execute the file because it bundles and runs with `ts-node` and is
able to parse the top-level typescript file. We could think about
converting the cypress config file back to Javascript, which would
anyway lose us the typing information (but allow the file to be
linted by the default Javascript configuration).

Bug: T372870
Change-Id: I0fbdeda0d9a800dc85d7810de38784f96e60110a