Streamline overly complex LexemeRdfBuilderTest
I can't really tell why the test setup was done like this, involving
a setUp function as well as managing state in class properties. This
is a rather trivial @dataProvider that feeds into rather trivial
mocks. Why make it so complicated? Maybe the plan was to add more
complicated test cases later?

This patch also replaces a type declaration in the LexemeRdfBuilder
with a more narrow interface. It appears like no code ever needs to
know the actual sub-type. Or does it?

Change-Id: I1ec0f5d6f71bfcc318b6d26f8e9643e957a226c2