Add strict types to data model classes
Normally, according to ADR 11 [1], we add strict types to files that
we’ve touched in a previous change, but in this case I think it makes
our lives easier if we establish a baseline of strict types in the core
data model, so that other code can rely on these methods returning the
right types. (This also means that PHPUnit mocks of these classes will
start to return the right types by default.)

Annoyingly, Phan apparently treats nullable return types in PHPdoc and
in real return types differently, so now a couple of places get warnings
because Lexeme::getId() returns a nullable LexemeId; add some asserts to
cover those.

Some unit tests for type errors can be removed now – there’s no point in
testing that the PHP engine enforces the declared types, IMHO. Also,
note that testGivenInvalidSerialization_unserializeThrowsAnException()
in FormIdTest and SenseIdTest was completely broken, because it
accidentally used $id as both the FormId/SenseId and the serialization,
so the construction would always fail because the “serialization” was a
FormId/SenseId instance rather than a string. Fix that by renaming half
the variable.

Also add strict types to the dummy objects and the data model tests to
match the data model itself. (But in the tests, I don’t bother to add
the return type `: void` to all test methods, IMHO that just bloats the
diff for no benefit. If we ever want to ensure that all test methods
return void, we can let phpcbf do that in one big commit later.)

[1]: https://doc.wikimedia.org/Wikibase/master/php/adr_0011.html

Change-Id: I41aee95642fb2809251a5ca712645147f0c910ea