Tolerate missing forms/senses in LexemePatcher
When users try to undo an edit (range) which affected a form or sense
that was since removed from the lexeme, it’s possible for the patcher to
see a diff involving form or sense IDs that can’t be found in the lexeme
to be patched. In this case, I think the correct thing is to just skip
those parts of the diff.

If the diff only contained those changes and nothing else, Wikibase will
report this as

> 1 change cannot be undone because the value has changed in the
meantime.
> There is nothing that can be undone here.

on the undo page, which sounds correct to me.

Note that this arguably restores the behavior the code previously had,
or at least intended to have: prior to changes Ib460ce4721 (commit
10b515c472) and Ia99f1b8832 (commit 781aa892c7), the LexemePatcher
skipped patching forms or senses if the $form or $sense was null; but
they can never be null. Indeed, prior to change I96b7f528b1 (commit
9d4624ca7c), the LexemePatcher even explicitly checked whether the form
existed before trying to get it (patching senses was apparently not
implemented at the time); if you just look at the change between the old
code of that commit and the new code of this commit, we’ve basically
just changed the code pattern from LBYL to EAFP (over the course of some
six years, and with a fair amount of confusion in between).

Bug: T284061
Bug: T326768
Change-Id: I486c36c29be36f92326857c95244d94e1b2373a1