Parallelize Cypress tests
parallelize tests using `cypress-parallel`

This also updates naming in `recentChanges` test and page object
to make it clear that all recent lexeme changes on the page are
considered, rather than just the most recent change. If it was
just the most recent change, this test could be flaky when tests
running in parallel create or modify lexemes. The conditions are
also tightened, looking for the expected values only in the element
containing the lexemeId. This has the bonus impact of making the
test slightly faster.

Bug: T374834
Change-Id: I05539e7ed7ad2ed04d808eae5815e871bdca6966