Skip tests to facilitate making Wikibase dataProviders static
non-static dataProvider functions are deprecated as of PHPUnit 10.
T337154 covers making these functions static in Wikibase. Some of
those functions are declared in an abstract class in Wikibase, but
implemented in this repo.

To facilitate the change, temporarily skip the affected tests. They
will be re-enabled in a followup commit

Bug: T337154
Change-Id: I9c2751b979cab2255cd9d88920663cdee17d3343