Make PHPUnit dataProvider methods static
Drop all the copies of invalidEntityProvider() and just hard-code an
Item as the invalid entity in each test – I don’t think it’s useful to
test these type checks with two different invalid entity types. (The
code being tested is just Assert::parameterType() anyway.) Also fix the
order of expectException() in ChangeOpLemmaEditTest (we shouldn’t call
it too soon).

The rest of the fixes are trivial.

Bug: T337151
Change-Id: I49e09a8086194a6431509f7b6c7a38c4d1650290