Remove $additionalLocalizedLanguages / wikibase-lexeme-language-name-*
All of these language names have been ported to the cldr extension and
are no longer necessary here. LexemeLanguageNameLookupFactory and
LexemeLanguageNameLookup are now effectively useless code and we should
replace them with normal Wikibase LanguageNameLookupFactory and
LanguageNameLookup respectively, but that can be done later.

The other files in language-names/ are not touched here – they should be
cleared out by l10n-bot as usual. After that, I assume we can remove the
whole directory.

Bug: T352922
Bug: T381370
Change-Id: Ifd91037f3f55e3d783d9f9758e13ad8bacf9afec
Depends-On: I7d610cbfb71fe4e007c1c63d136b713516e0e4b3