build: Disable MediaWiki.WhiteSpace.SpaceBeforeSingleLineComment sniff
There are 37 files with 86 findings.

This extension has many comments after code lines.
Sniffs listed in <exclude> get autofixed by libup and indicate that the
sniff should be fixed in the long term.
Disable the sniff with serverity = 0 to indicate that
this sniff does not need fixing in this extension.

Change-Id: Ia100ea2da60607a302f779f7040b4296307366a2