docs: Fix broken documentation links in Doxygen
The automatically generated reference names change
between doxygen release and are thus generally not
to be relied upon.

Tested with various doxygen versions between 1.9.4
and the current 1.13.2.

This is similar to c4ec3483e77b in Wikibase, but not
using a special "FILTER_PATTERNS" script, but rather
just explicitly declaring the page references for all
files.

Bug: T355569
Change-Id: I74b17b6a25c16999bb2f2be9be6f8647576c87d1