Use OPT_LANG instead of WikibaseRepo.UserLanguage service
The UserLanguage service is deprecated (for good reason). Instead, it
appears that OPT_LANG is always set in the formatter options that we’re
passed, so we can and should use that instead. In one place where we
need a fallback chain, we can even get that from a separate option.

Bug: T281799
Change-Id: Ia28e150ee738200e1f5ed85b52fcafbf37e74b1b