Use baserevid correctly in wbladdform+wbladdsense
The edit should be based on the revision specified in the baserevid
parameter; applying that to the latest revision is EditEntity’s
responsibility. (Compare other API modules, including wblremoveform and
wbleditsenseelements, which already do it correctly.)

I *think* the change from wikibase-self-conflict-patched to
wikibase-conflict-patched in the tests is correct. The conflict is
technically a self-conflict (the lexeme is created by a random TestUser
because saveEntity() uses $this->getTestUser(), but both API requests
are made by UTSysop because doApiRequest() uses $this->getTestSysop()),
but the self-conflict message is only used for conflicts that could not
be patched cleanly, and where Wikibase would have refused to save the
edit at all if it wasn’t a self-conflict (the message says “overriding
some of your own intermediate changes”).

Bug: T392372
Change-Id: Ic4cc78894c18155ddf41542364ed1c3c41201ff9