Skip restoring already-restored form/sense
(The rest of this commit message discusses only forms WLOG.)

If LexemePatcher encounters a ConflictException while trying to add a
form, there are two possible cases: the conflicting form is “the same”
form, or it’s an unrelated form that happened to get assigned the same
sequential ID. The second case can happen if the wbladdform API is used
with a baserevid and another form was added since the base revision: in
this case, we don’t want the edit to be saved, as the API caller might
be trying to add the same form that was already saved in the other edit
since the baserevid. The first case can, I think, only happen when
undoing a form deletion, and more specifically the ConflictException can
only happen if the form deletion was already undone in the meantime: in
this case, we want the patcher to simply skip the form addition, and the
overall edit will probably report “nothing left to do”.

We can distinguish between these two cases based on the next form ID
diff in the patch. The next form ID is increased whenever a new form is
added; however, it is not decreased when a form is removed, and
consequently the diff for restoring a form does not affect the next form
ID either. So if the next form ID diff is empty, we want to ignore the
ConflictException; otherwise, (re)throw it as an indication of the
genuine edit conflict.

I extracted the patchAddForm() and patchAddSense() methods purely to
make the Generic.Metrics.NestingLevel.TooHigh phpcs sniff happy; I’m
open to other ways to structure this code.

In the tests, the submit() method is based on Wikibase’s ActionTestCase
(specifically createAction() and callAction()). However, we do not
extend ActionTestCase directly, because that class also creates a bunch
of test data, and by doing this in addDBDataOnce() also has the
unpleasant side-effect of not clearing the page etc. tables between test
runs (which means that the separate tests added here, each creating a
new Lexeme:L1 page, would conflict with one another).

Bug: T392372
Change-Id: I65f26a50fc7a136a2376bb94d08e5942872c4173