fix: Fix styling of anon edit warning boxes
Special:MergeLexemes was completely unstyled. Special:NewLexeme already
had a styling that resembled a Codex message box, but it was not quite
the same: it had a black icon instead of a brown one, its width was
flexible instead of 100% (note that this is visible only with an older
skin like Vector 2010, or with temp accounts enabled, not by default on
Wikidata). For consistency with both other pages and the JS version of
the same page (which already uses Codex), use `Html::warningBox()` here
as well, and remove the custom styling.

Bug: T404500
Change-Id: I6ad0d72c4dc525ab6fec2323975e67058786fd54